SuspiciousGazeDetection/requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
albumentations>=1.3.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Object Detection & Tracking
ultralytics>=8.0.0  # YOLOv8
boxmot>=10.0.0      # StrongSORT, DeepSORT, etc.

# Head Pose Estimation
onnx>=1.14.0
onnxruntime-gpu>=1.15.0

# Neural Network Components
einops>=0.6.0       # Tensor operations
timm>=0.9.0         # Vision models

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Logging & Experiment Tracking
wandb>=0.15.0
loguru>=0.7.0

# Utilities
pillow>=9.5.0
requests>=2.31.0
python-dotenv>=1.0.0

# Video Processing
decord>=0.6.0
av>=10.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
